+++ src/types/stats.ts
@@ -1,0 +1,17 @@
+export interface ParticipantStats {
+  team: number;
+  study: number;
+  project: number;
+  jobPrep: number;
+  etc: number;
+}
+
+export interface BasicProfileStats {
+  postCount: number;
+  meetingCount: number;
+  favoriteCount: number;
+}
+
+export interface CreatedMeetingSummary {
+  category: "PROJECT" | "STUDY";
+}
